Improve naming tests (#2094)
* Fix test names

* Add missing test in EnumNamingSpec

* Split ParameterNamingSpec

* Don't use NamingRules for the tests

* Remove the cases that are already tested

* Add more annotations Language("kotlin")

* Use compileAndLint

* Use VariableNaming instead of NamingRules

* Extract VariableNamingSpec from NamingRulesSpec
