+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/naming/PackageNamingSpec.kt
@@ -10,19 +10,19 @@ class PackageNamingSpec : Spek({
     describe("PackageNaming rule") {
 
         it("should find a uppercase package name") {
-            assertThat(NamingRules().compileAndLint("package FOO.BAR")).hasSize(1)
+            assertThat(PackageNaming().compileAndLint("package FOO.BAR")).hasSize(1)
         }
 
         it("should find a upper camel case package name") {
-            assertThat(NamingRules().compileAndLint("package Foo.Bar")).hasSize(1)
+            assertThat(PackageNaming().compileAndLint("package Foo.Bar")).hasSize(1)
         }
 
         it("should find a camel case package name") {
-            assertThat(NamingRules().compileAndLint("package fOO.bAR")).hasSize(1)
+            assertThat(PackageNaming().compileAndLint("package fOO.bAR")).hasSize(1)
         }
 
         it("should check an valid package name") {
-            assertThat(NamingRules().compileAndLint("package foo.bar")).isEmpty()
+            assertThat(PackageNaming().compileAndLint("package foo.bar")).isEmpty()
         }
     }
 })
